RF64 file bug fix
Resolved bug where in RF64 files, the incorrect subchunk file size (of the AXML subchunk) was being calculated after the much larger DS64 subchunk size was calculated. Because the size was being read in to via pointer, MSBs were not reset and the calcualted size was much larger than it should have been. Reinstated line of code where the subchunk_size was reset to 0 prior to reading in subchunk size. Also, implemented split and mini-chunk parsing of large AXML subchunk.